dynamic map loading!!!

+++ Astar/src/aStar/MapV2.java
@@ -10,7 +10,8 @@ import java.util.Scanner;
 public class MapV2 {
 
     private int size_x, size_y;
-    private Node[][] loadedRegion;
+    private Node[][] loadedNodes;
+    private MapRegion loadedRegion = null;
     private File map;
 
     /**
@@ -22,21 +23,27 @@ public class MapV2 {
      */
     public MapV2(File file, int size_x, int size_y, MapRegion region){
         map = file; this.size_x = size_x; this.size_y = size_y;
-        try {
-            //System.out.println("serializing map");
-            PrintWriter writer = new PrintWriter(new BufferedWriter(new FileWriter(file)));
-            for(int r = 0; r < size_y; r++){
-                String toWrite = "";
-                for(int c = 0; c < size_x; c++){
-                    toWrite += Node.TraversalState.DEFAULT + "\t"; //TSV
+        //System.out.println("set class vars");
+        //System.out.println(map.isFile() + "");
+        if(map.isFile()) {
+            try {
+                System.out.println("writing blank map");
+                PrintWriter writer = new PrintWriter(new BufferedWriter(new FileWriter(file)));
+                for (int r = 0; r < size_y; r++) {
+                    String toWrite = "";
+                    for (int c = 0; c < size_x; c++) {
+                        toWrite += Node.TraversalState.DEFAULT + "\t"; //TSV
+                    }
+                    writer.println(toWrite);
+                    writer.flush();
                 }
-                writer.println(toWrite);
-                writer.flush();
-            }
-            loadRegion(region);
 
-        } catch (IOException e) {
-            System.out.println("failed to serialize map to file");
+
+            } catch (IOException e) {
+                System.out.println("failed to serialize map to file");
+            }
+            //System.out.println("attempting to load " + region);
+            loadRegion(region);
         }
     }
 
@@ -51,24 +58,28 @@ public class MapV2 {
     }
 
     public void loadRegion (MapRegion region){
-        writeRegion(region);
-        //System.out.println("deserializing map");
+        writeRegion(loadedRegion);
+        //System.out.println("called to load " + region);
         try {
             Scanner scanner = new Scanner(map);
-            loadedRegion = new Node[region.y_size][region.x_size];
+            loadedNodes = new Node[region.y_size][region.x_size];
+            //System.out.println("created array");
             for(int r = 0; r < region.y_off + region.y_size; r++){
+                //System.out.println("row " + r);
                 if(r <  region.y_off){
                     scanner.nextLine();
                     continue;
                 }
                 for(int c = 0; c < region.x_off + region.x_size; c++){
-                    if(r < region.x_off){
+                    if(c < region.x_off){
                         scanner.nextInt();
                         continue;
                     }
-                    loadedRegion[r][c] = new Node(r,c,Node.getStateFromInt(scanner.nextInt()));
+                    loadedNodes[r - region.y_off][c - region.x_off] = new Node(r,c,Node.getStateFromInt(scanner.nextInt()));
                 }
             }
+            loadedRegion = region;
+            System.out.println("loaded " + region);
         } catch (FileNotFoundException e) {
             System.out.println("failed to de-serialize region");
         }
@@ -76,23 +87,25 @@ public class MapV2 {
     }
 
     private void writeRegion(MapRegion region){
+        //System.out.println("writing " + loadedRegion);
+        if(region == null) return;
         try {
             List<String> lines = Files.readAllLines(map.toPath());
-            for(int r = 0; r < region.y_off + region.y_size; r++){
-                if(r <  region.y_off){
-                    continue;
+            //System.out.println("Read file contents");
+            for(int r = region.y_off; r < region.y_off + region.y_size; r++){
+                String out = "";
+                for(int c = region.x_off; c < region.x_off + region.x_size; c++){
+                    out += loadedNodes[r - region.y_off][c - region.x_off].serialize() + "\t";
                 }
-                for(int c = 0; c < region.x_off + region.x_size; c++){
-                    if(r < region.x_off){
-                        scanner.nextInt();
-                        continue;
-                    }
-                    //exec
-                }
+                //System.out.println("re-writing line " + r);
+                out = ((region.x_off > 0) ? lines.get(r).substring(0, region.x_off * 2) : "") + out +
+                        ((lines.get(r).length() > (region.x_off + region.x_size) * 2)? lines.get(r).substring((region.x_off + region.x_size + 1) * 2) : "");
+                lines.set(r, out);
             }
         } catch (IOException e) {
             System.out.println("failed to serialize region");
         }
+        System.out.println(region + " written");
     }
 
     /**
@@ -104,7 +117,7 @@ public class MapV2 {
      */
     public Node getNode(int x, int y){
         try {
-            return loadedRegion[y][x];
+            return loadedNodes[y][x];
         }
         catch (ArrayIndexOutOfBoundsException e ){
             return null;
@@ -112,7 +125,7 @@ public class MapV2 {
     }
 
     public void close(){
-        loadedRegion = null;
+        loadedNodes = null;
         System.gc();
     }
 
@@ -121,26 +134,26 @@ public class MapV2 {
         LinkedList<Node> toReturn = new LinkedList<>();
 
         //left
-        if(node.getPos_x() - 1 > -1 && loadedRegion[node.getPos_y()][node.getPos_x() - 1].canTraverse() && !loadedRegion[node.getPos_y()][node.getPos_x() - 1].getSearched()){
-            loadedRegion[node.getPos_y()][node.getPos_x() - 1].setCameFrom(node);
-            toReturn.add(loadedRegion[node.getPos_y()][node.getPos_x() - 1]);
+        if(node.getPos_x() - 1 > -1 && loadedNodes[node.getPos_y()][node.getPos_x() - 1].canTraverse() && !loadedNodes[node.getPos_y()][node.getPos_x() - 1].getSearched()){
+            loadedNodes[node.getPos_y()][node.getPos_x() - 1].setCameFrom(node);
+            toReturn.add(loadedNodes[node.getPos_y()][node.getPos_x() - 1]);
         }
 
         //right
-        if(node.getPos_x() + 1 < loadedRegion[0].length && loadedRegion[node.getPos_y()][node.getPos_x() + 1].canTraverse() && !loadedRegion[node.getPos_y()][node.getPos_x() + 1].getSearched()) {
-            loadedRegion[node.getPos_y()][node.getPos_x() + 1].setCameFrom(node);
-            toReturn.add(loadedRegion[node.getPos_y()][node.getPos_x() + 1]);
+        if(node.getPos_x() + 1 < loadedNodes[0].length && loadedNodes[node.getPos_y()][node.getPos_x() + 1].canTraverse() && !loadedNodes[node.getPos_y()][node.getPos_x() + 1].getSearched()) {
+            loadedNodes[node.getPos_y()][node.getPos_x() + 1].setCameFrom(node);
+            toReturn.add(loadedNodes[node.getPos_y()][node.getPos_x() + 1]);
         }
 
         //above
-        if(node.getPos_y() - 1 > -1 && loadedRegion[node.getPos_y() - 1][node.getPos_x()].canTraverse() && !loadedRegion[node.getPos_y() - 1][node.getPos_x()].getSearched()) {
-            loadedRegion[node.getPos_y() - 1][node.getPos_x()].setCameFrom(node);
-            toReturn.add(loadedRegion[node.getPos_y() - 1][node.getPos_x()]);
+        if(node.getPos_y() - 1 > -1 && loadedNodes[node.getPos_y() - 1][node.getPos_x()].canTraverse() && !loadedNodes[node.getPos_y() - 1][node.getPos_x()].getSearched()) {
+            loadedNodes[node.getPos_y() - 1][node.getPos_x()].setCameFrom(node);
+            toReturn.add(loadedNodes[node.getPos_y() - 1][node.getPos_x()]);
         }
         //below
-        if(node.getPos_y() + 1 < loadedRegion.length && loadedRegion[node.getPos_y() + 1][node.getPos_x()].canTraverse() && !loadedRegion[node.getPos_y() + 1][node.getPos_x()].getSearched()) {
-            loadedRegion[node.getPos_y() + 1][node.getPos_x()].setCameFrom(node);
-            toReturn.add(loadedRegion[node.getPos_y() + 1][node.getPos_x()]);
+        if(node.getPos_y() + 1 < loadedNodes.length && loadedNodes[node.getPos_y() + 1][node.getPos_x()].canTraverse() && !loadedNodes[node.getPos_y() + 1][node.getPos_x()].getSearched()) {
+            loadedNodes[node.getPos_y() + 1][node.getPos_x()].setCameFrom(node);
+            toReturn.add(loadedNodes[node.getPos_y() + 1][node.getPos_x()]);
         }
         return toReturn;
     }
@@ -173,7 +186,7 @@ public class MapV2 {
     }
 
     private void reset(){
-        for(Node[] nodes : loadedRegion){
+        for(Node[] nodes : loadedNodes){
             for(Node node : nodes){
                 node.reset();
             }
@@ -182,7 +195,7 @@ public class MapV2 {
 
     private void setGoal(Node goal){
         //System.out.println("setting goal");
-        for(Node[] nodes : loadedRegion){
+        for(Node[] nodes : loadedNodes){
             for(Node node : nodes){
                 node.setGoal(goal);
             }
